Extract `addQuickSubmit()` as separate feature
