+++ source/features/allow-submission-via-ctrl-enter.tsx
@@ -1,0 +1,19 @@
+import select from 'select-dom';
+import * as pageDetect from 'github-url-detection';
+
+import features from '.';
+
+function addQuickSubmit(): void {
+	select('input#commit-summary-input')!.classList.add('js-quick-submit');
+}
+
+void features.add(import.meta.url, {
+	shortcuts: {
+		'ctrl enter': 'Publish a new/edited file',
+	},
+	include: [
+		pageDetect.isNewFile,
+		pageDetect.isEditingFile,
+	],
+	init: addQuickSubmit,
+});
